+++ src/Enqueue/Blocks/EnqueueBlocks.php
@@ -20,7 +20,6 @@ use EightshiftForms\CustomPostType\Result;
 use EightshiftForms\CustomPostType\Forms;
 use EightshiftForms\Enqueue\SharedEnqueue;
 use EightshiftForms\Enqueue\Captcha\EnqueueCaptcha;
-use EightshiftForms\General\SettingsGeneral;
 use EightshiftForms\Geolocation\SettingsGeolocation;
 use EightshiftFormsVendor\EightshiftFormsUtils\Helpers\UtilsIntegrationsHelper;
 use EightshiftForms\Hooks\FiltersOuputMock;
@@ -156,7 +155,6 @@ class EnqueueBlocks extends AbstractEnqueueBlocks
 		$additionalBlocksFilterName = UtilsHooksHelper::getFilterName(['blocks', 'additionalBlocks']);
 		$formsStyleOptionsFilterName = UtilsHooksHelper::getFilterName(['block', 'forms', 'styleOptions']);
 		$formsUseCustomResultOutputFeatureFilterName = UtilsHooksHelper::getFilterName(['block', 'forms', 'useCustomResultOutputFeature']);
-		$formsUseLegacyTnxPageFeatureFilterName = UtilsHooksHelper::getFilterName(['block', 'forms', 'useLegacyTnxPageFeature']);
 		$fieldStyleOptionsFilterName = UtilsHooksHelper::getFilterName(['block', 'field', 'styleOptions']);
 		$breakpointsFilterName = UtilsHooksHelper::getFilterName(['blocks', 'mediaBreakpoints']);
 		$formSelectorTemplatesFilterName = UtilsHooksHelper::getFilterName(['block', 'formSelector', 'formTemplates']);
@@ -164,7 +162,6 @@ class EnqueueBlocks extends AbstractEnqueueBlocks
 		$output['additionalBlocks'] = \apply_filters($additionalBlocksFilterName, []);
 		$output['formsBlockStyleOptions'] = \apply_filters($formsStyleOptionsFilterName, []);
 		$output['formsUseCustomResultOutputFeature'] = \apply_filters($formsUseCustomResultOutputFeatureFilterName, false);
-		$output['formsUseLegacyTnxPageFeature'] = \apply_filters($formsUseLegacyTnxPageFeatureFilterName, false);
 		$output['fieldBlockStyleOptions'] = \apply_filters($fieldStyleOptionsFilterName, []);
 		$output['validationPatternsOptions'] = ValidationPatterns::getValidationPatternsEditor();
 		$output['mediaBreakpoints'] = \apply_filters($breakpointsFilterName, []);
@@ -173,11 +170,6 @@ class EnqueueBlocks extends AbstractEnqueueBlocks
 		$output['postTypes'] = [
 			'results' => Result::POST_TYPE_SLUG,
 			'forms' => Forms::POST_TYPE_SLUG,
-		];
-
-		// Legacy.
-		$output['settings'] = [
-			'successRedirectVariations' => UtilsSettingsHelper::getOptionValueGroup(SettingsGeneral::SETTINGS_GENERAL_SUCCESS_REDIRECT_VARIATION_OPTIONS_KEY),
 		];
 
 		$output['use'] = [