+++ src/General/SettingsGeneral.php
@@ -18,23 +18,17 @@ use EightshiftFormsVendor\EightshiftFormsUtils\Helpers\UtilsSettingsHelper;
 use EightshiftForms\Hooks\FiltersOuputMock;
 use EightshiftFormsVendor\EightshiftFormsUtils\Config\UtilsConfig;
 use EightshiftFormsVendor\EightshiftFormsUtils\Helpers\UtilsHelper;
-use EightshiftFormsVendor\EightshiftFormsUtils\Settings\UtilsSettingGlobalInterface;
 use EightshiftFormsVendor\EightshiftLibs\Services\ServiceInterface;
 
 /**
  * SettingsGeneral class.
  */
-class SettingsGeneral implements UtilsSettingGlobalInterface, UtilsSettingInterface, ServiceInterface
+class SettingsGeneral implements UtilsSettingInterface, ServiceInterface
 {
 	/**
 	 * Filter settings key.
 	 */
 	public const FILTER_SETTINGS_NAME = 'es_forms_settings_general';
-
-	/**
-	 * Filter global settings key.
-	 */
-	public const FILTER_SETTINGS_GLOBAL_NAME = 'es_forms_settings_global_general';
 
 	/**
 	 * Settings key.
@@ -94,20 +88,6 @@ class SettingsGeneral implements UtilsSettingGlobalInterface, UtilsSettingInterf
 	public const SETTINGS_HIDE_FORM_ON_SUCCESS_KEY = 'hide-form-on-success';
 
 	/**
-	 * General success redirect variation options key.
-	 *
-	 * @deprecated 5.0.0 It will be removed in the 6.0.0 release.
-	 */
-	public const SETTINGS_GENERAL_SUCCESS_REDIRECT_VARIATION_OPTIONS_KEY = 'general-success-redirect-variation-options';
-
-	/**
-	 * General success redirect variation key.
-	 *
-	 * @deprecated 5.0.0 It will be removed in the 6.0.0 release.
-	 */
-	public const SETTINGS_GENERAL_SUCCESS_REDIRECT_VARIATION_KEY = 'general-success-redirect-variation';
-
-	/**
 	 * Increment meta key.
 	 *
 	 * @var string
@@ -132,7 +112,6 @@ class SettingsGeneral implements UtilsSettingGlobalInterface, UtilsSettingInterf
 	public function register(): void
 	{
 		\add_filter(self::FILTER_SETTINGS_NAME, [$this, 'getSettingsData']);
-		\add_filter(self::FILTER_SETTINGS_GLOBAL_NAME, [$this, 'getSettingsGlobalData']);
 	}
 
 	/**
@@ -159,8 +138,6 @@ class SettingsGeneral implements UtilsSettingGlobalInterface, UtilsSettingInterf
 		$variation = FiltersOuputMock::getVariationFilterValue($formType, $formId, []);
 		$trackingEventName = FiltersOuputMock::getTrackingEventNameFilterValue($formType, $formId);
 		$trackingAdditionalData = FiltersOuputMock::getTrackingAditionalDataFilterValue($formType, $formId);
-
-		$successRedirectVariation = UtilsSettingsHelper::getSettingValue(SettingsGeneral::SETTINGS_GENERAL_SUCCESS_REDIRECT_VARIATION_KEY, $formId);
 
 		return [
 			UtilsSettingsOutputHelper::getIntro(self::SETTINGS_TYPE_KEY),
@@ -257,31 +234,6 @@ class SettingsGeneral implements UtilsSettingGlobalInterface, UtilsSettingInterf
 										'checkboxAsToggle' => true,
 									],
 								],
-							],
-							[
-								'component' => 'divider',
-								'dividerExtraVSpacing' => true,
-							],
-							[
-								'component' => 'select',
-								'selectFieldLabel' => \__('Redirect variation - Deprecated', 'eightshift-forms'),
-								'selectName' => UtilsSettingsHelper::getSettingName(self::SETTINGS_GENERAL_SUCCESS_REDIRECT_VARIATION_KEY),
-								'selectPlaceholder' => \__('Pick an option', 'eightshift-forms'),
-								// translators: %s will be replaced with forms field name and filter output copy.
-								'selectFieldHelp' => \sprintf(\__('Variation value will be appended to the redirect URL as a parameter to allow you to provide different version of the "Thank you" page.', 'eightshift-forms')),
-								'selectContent' => \array_values(
-									\array_map(
-										function ($selectOption) use ($successRedirectVariation) {
-											return [
-												'component' => 'select-option',
-												'selectOptionLabel' => $selectOption[1],
-												'selectOptionValue' => $selectOption[0],
-												'selectOptionIsSelected' => $selectOption[0] === $successRedirectVariation,
-											];
-										},
-										UtilsSettingsHelper::getOptionValueGroup(SettingsGeneral::SETTINGS_GENERAL_SUCCESS_REDIRECT_VARIATION_OPTIONS_KEY)
-									)
-								),
 							],
 						],
 					],
@@ -483,41 +435,6 @@ class SettingsGeneral implements UtilsSettingGlobalInterface, UtilsSettingInterf
 						],
 					],
 				]
-			],
-		];
-	}
-
-	/**
-	 * Get global settings array for building settings page.
-	 *
-	 * @return array<int, array<string, mixed>>
-	 */
-	public function getSettingsGlobalData(): array
-	{
-		return [
-			UtilsSettingsOutputHelper::getIntro(self::SETTINGS_TYPE_KEY),
-			[
-				'component' => 'tabs',
-				'tabsContent' => [
-					[
-						'component' => 'tab',
-						'tabLabel' => \__('Submit', 'eightshift-forms'),
-						'tabContent' => [
-							[
-								'component' => 'textarea',
-								'textareaName' => UtilsSettingsHelper::getOptionName(self::SETTINGS_GENERAL_SUCCESS_REDIRECT_VARIATION_OPTIONS_KEY),
-								'textareaIsMonospace' => true,
-								'textareaSaveAsJson' => true,
-								'textareaFieldLabel' => \__('After form submission redirect variants - Deprecated', 'eightshift-forms'),
-								// translators: %s will be replaced with local validation patterns.
-								'textareaFieldHelp' => UtilsGeneralHelper::minifyString(\sprintf(\__("
-									These entries will populate the dropdown in the \"Thank you page\" options so different styles can be selected based on the use case.<br /><br />
-									Provide one key-value pair per line, in this format: %1\$s", 'eightshift-forms'), '<code>slug : label</code>')),
-								'textareaValue' => UtilsSettingsHelper::getOptionValueAsJson(self::SETTINGS_GENERAL_SUCCESS_REDIRECT_VARIATION_OPTIONS_KEY, 2),
-							],
-						],
-					],
-				],
 			],
 		];
 	}