+++ CHANGELOG.md
@@ -4,11 +4,20 @@ All notable changes to this project will be documented in this file.
 
 This projects adheres to [Semantic Versioning](https://semver.org/) and [Keep a CHANGELOG](https://keepachangelog.com/).
 
-## [5.9.11]
+## [6.0.0]
 
 ### Fixed
 
 - multiple integrations for new fields types.
+
+### Removed
+
+- Legacy variation data passed to the success redirect url.
+- `useLegacyTnxPageFeature` filter.
+
+### Added
+
+- New helper function for getting the field value `getParamValue`.
 
 ## [5.9.10]
 
@@ -1059,7 +1068,7 @@ This projects adheres to [Semantic Versioning](https://semver.org/) and [Keep a 
 
 - Initial production release.
 
-[5.9.11]: https://github.com/infinum/eightshift-forms/compare/5.9.10...5.9.11
+[6.0.0]: https://github.com/infinum/eightshift-forms/compare/5.9.10...6.0.0
 [5.9.10]: https://github.com/infinum/eightshift-forms/compare/5.9.9...5.9.10
 [5.9.9]: https://github.com/infinum/eightshift-forms/compare/5.9.8...5.9.9
 [5.9.8]: https://github.com/infinum/eightshift-forms/compare/5.9.7...5.9.8